Add support for dynamic block splitting to actor pool.
