+++ python/ray/data/_internal/execution/operators/map_task_submitter.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
-from typing import List, Union, Tuple
-from ray.data.block import Block, BlockMetadata
+from typing import List, Union, Tuple, Callable, Iterator
+from ray.data.block import Block, BlockAccessor, BlockMetadata, BlockExecStats
 from ray.types import ObjectRef
 from ray._raylet import ObjectRefGenerator
 
@@ -58,3 +58,30 @@ class MapTaskSubmitter(ABC):
             task_refs: The output refs for all of the tasks submitted by this submitter.
         """
         raise NotImplementedError
+
+
+def _map_task(
+    fn: Callable[[Iterator[Block]], Iterator[Block]],
+    *blocks: Block,
+) -> Iterator[Union[Block, List[BlockMetadata]]]:
+    """Remote function for a single operator task.
+
+    Args:
+        fn: The callable that takes Iterator[Block] as input and returns
+            Iterator[Block] as output.
+        blocks: The concrete block values from the task ref bundle.
+
+    Returns:
+        A generator of blocks, followed by the list of BlockMetadata for the blocks
+        as the last generator return.
+    """
+    output_metadata = []
+    stats = BlockExecStats.builder()
+    for b_out in fn(iter(blocks)):
+        # TODO(Clark): Add input file propagation from input blocks.
+        m_out = BlockAccessor.for_block(b_out).get_metadata([], None)
+        m_out.exec_stats = stats.build()
+        output_metadata.append(m_out)
+        yield b_out
+        stats = BlockExecStats.builder()
+    yield output_metadata
