+++ python/ray/data/_internal/execution/operators/task_pool_submitter.py
@@ -1,8 +1,11 @@
 from typing import Dict, Any, Iterator, Callable, Union, List
 
 import ray
-from ray.data.block import Block, BlockAccessor, BlockMetadata, BlockExecStats
-from ray.data._internal.execution.operators.map_task_submitter import MapTaskSubmitter
+from ray.data.block import Block
+from ray.data._internal.execution.operators.map_task_submitter import (
+    MapTaskSubmitter,
+    _map_task,
+)
 from ray.data._internal.remote_fn import cached_remote_fn
 from ray.types import ObjectRef
 from ray._raylet import ObjectRefGenerator
@@ -48,30 +51,3 @@ class TaskPoolSubmitter(MapTaskSubmitter):
                 # a different error, or cancellation failed. In all cases, we
                 # swallow the exception.
                 pass
-
-
-def _map_task(
-    fn: Callable[[Iterator[Block]], Iterator[Block]],
-    *blocks: Block,
-) -> Iterator[Union[Block, List[BlockMetadata]]]:
-    """Remote function for a single operator task.
-
-    Args:
-        fn: The callable that takes Iterator[Block] as input and returns
-            Iterator[Block] as output.
-        blocks: The concrete block values from the task ref bundle.
-
-    Returns:
-        A generator of blocks, followed by the list of BlockMetadata for the blocks
-        as the last generator return.
-    """
-    output_metadata = []
-    stats = BlockExecStats.builder()
-    for b_out in fn(iter(blocks)):
-        # TODO(Clark): Add input file propagation from input blocks.
-        m_out = BlockAccessor.for_block(b_out).get_metadata([], None)
-        m_out.exec_stats = stats.build()
-        output_metadata.append(m_out)
-        yield b_out
-        stats = BlockExecStats.builder()
-    yield output_metadata
